Correctly teleport to highest block instead of spawn (fixes #178)
The fix for issue #57 added some limits to the values searched when looking for
a safe Y coordinate for the teleport destination, including calling
getHighestBlockYAt() to determine the Y coordinate of the highest non-empty
block.  In non-nether cases with normal terrain that doesn't approach the
world's maximum Y values, highestBlockBoundary will be set to 1 greater than the
return value of getHighestBlockYAt(), i.e. the Y coordinate of the empty block
above the highest non-empty block.

In the non-nether, non-flying case, limTop was previously set to this value, and
the safe Y coordinate search would proceed up to *but not including* this limit,
so the maximum Y coordinate searched to check if it was empty would be the
highest *non*-empty y coordinate.  However, the player's current Y coordinate is
checked first even if it is equal to limTop (unless it is somehow not greater
than limBot).  As a result, if the only safe block for the player to be
teleported on top of was getHighestBlockYAt(), they would only be teleported
there if it wouldn't result in a change in their Y coordinate, otherwise they
would be teleported to spawn.

This fix expands the search to include highestBlockBoundary.

highestBlockBoundary cannot be greater than world.getMaxHeight() - 2, so the
search for a Y coordinate will now include that value rather than excluding it.
isSafeSpot() only adds 1 to that value, giving world.getMaxHeight() - 1, which
is still within the valid Y coordinate range.

This might also fix issue #115.