Allow teleporting player partially or fully above the world
For worlds other than the nether (which is a special case due to the
bedrock ceiling) the maximum Y coordinate to which a player would be
teleported was previously world.getMaxHeight() - 2.  This enabled
isSafeSpot() to check the blocks at Y and Y + 1 to ensure that they
were breathable.  However, this meant that if there were blocks all
the way up to 1 or 2 less than world.getMaxHeight(), the player would
never be teleported on top of them, and instead would be teleported to
an air gap somewhere below them, or to spawn if no such gap was found.

This commit enables the player to be teleported to Y coordinates up to
world.getMaxHeight() so that they can be teleported on top of the
highest possible blocks (which would be at world.getMaxHeight() - 1).
As part of this, isSafeSpot() is updated to treat the void at Y >=
world.getMaxHeight() as breathable.